Test long searches
- Split test_browse into test_browse and test_search
- Added code to retry searches that timeout
- Added test using a long search string
